+++ app/navigation/ElderNavigator.tsx
@@ -2,10 +2,10 @@ import { createNativeStackNavigator } from "@react-navigation/native-stack";
 import ElderTabNavigator from "./ElderTabNavigator";
 import SharedHealthStackNavigator from "./SharedHealthStackNavigator";
 import { ElderTabParamList } from "./navigation";
-import AssignCaregiverScreen from "../elder_only/assigncaregiver";
 import ElderPerimeterMap from "../elder_only/MapElder";
-import RemoveCaregiver from "../elder_only/removecaregiver";
-import ViewArduino from "../elder_only/viewarduino";
+import AssignCaregiverScreen from "../elder_only/AssignCaregiver";
+import RemoveCaregiver from "../elder_only/RemoveCaregiver";
+import ViewArduino from "../elder_only/ViewArduino";
 
 const Stack = createNativeStackNavigator<ElderTabParamList>();
