+++ providers/AuthenticationProvider.tsx
@@ -1,6 +1,5 @@
 import { createContext, useCallback, useContext, useEffect, useState } from "react";
 import { createUserRequest } from "@/apis/registerAPI";
-import { RegisterForm } from "@/app/Register";
 import { LoginResponse, loginUserRequest } from "@/apis/loginAPI";
 import { setBearer } from "@/apis/axiosConfig";
 import * as SecureStore from "expo-secure-store";
@@ -8,7 +7,8 @@ import { useToast } from "./ToastProvider";
 import { addLogoutListener, removeLogoutListener } from "@/utils/logoutEmitter";
 import { revokeRefreshTokenAPI } from "@/apis/revokeRefreshTokenAPI";
 import { useNavigationContainerRef } from "@react-navigation/native";
-import { LoginForm } from "@/app/Login";
+import { RegisterForm } from "@/app/auth/Register";
+import { LoginForm } from "@/app/auth/Login";
 
 /**
  * Represents a potentially authenticated user and provides authentication-related actions.
